+++ geotrek/trekking/models.py
@@ -7,6 +7,7 @@ from django.conf import settings
 from django.contrib.gis.db import models
 from django.core.urlresolvers import reverse
 from django.utils.translation import ugettext_lazy as _
+from django.utils.html import strip_tags
 from django.template.defaultfilters import slugify
 
 from easy_thumbnails.alias import aliases
@@ -167,8 +168,9 @@ class Trek(PicturesMixin, MapEntityMixin, Topology):
     related_treks = models.ManyToManyField('self', through='TrekRelationship',
                                            verbose_name=_(u"Related treks"), symmetrical=False,
                                            related_name='related_treks+')  # Hide reverse attribute
-    information_desk = models.ForeignKey('InformationDesk', related_name='treks',
-                                         blank=True, null=True, verbose_name=_(u"Information Desk"), db_column='renseignement')
+    information_desks = models.ManyToManyField('InformationDesk',
+                                               db_table="o_r_itineraire_renseignement", blank=True, null=True,
+                                               verbose_name=_(u"Information desks"))
 
     objects = Topology.get_manager_cls(models.GeoManager)()
 
@@ -321,14 +323,6 @@ class Trek(PicturesMixin, MapEntityMixin, Topology):
         return {'id': self.difficulty.pk,
                 'pictogram': pictogram,
                 'label': self.difficulty.difficulty}
-
-    @property
-    def serializable_information_desk(self):
-        if not self.information_desk:
-            return None
-        return {'id': self.information_desk.pk,
-                'name': self.information_desk.name,
-                'description': self.information_desk.description}
 
     @property
     def serializable_themes(self):
@@ -689,6 +683,26 @@ class InformationDesk(models.Model):
     name = models.CharField(verbose_name=_(u"Title"), max_length=256, db_column='nom')
     description = models.TextField(verbose_name=_(u"Description"), blank=True, db_column='description',
                                    help_text=_(u"Brief description"))
+    phone = models.CharField(verbose_name=_(u"Phone"), max_length=32,
+                             blank=True, null=True, db_column='telephone')
+    email = models.EmailField(verbose_name=_(u"Email"), max_length=256, db_column='email',
+                              blank=True, null=True)
+    website = models.URLField(verbose_name=_(u"Website"), max_length=256, db_column='website',
+                              blank=True, null=True)
+    photo = models.FileField(verbose_name=_(u"Photo"), upload_to=settings.UPLOAD_DIR,
+                             db_column='photo', max_length=512, blank=True, null=True)
+
+    street = models.CharField(verbose_name=_(u"Street"), max_length=256,
+                              blank=True, null=True, db_column='rue')
+    postal_code = models.IntegerField(verbose_name=_(u"Postal code"),
+                                      blank=True, null=True, db_column='code')
+    municipality = models.CharField(verbose_name=_(u"Municipality"),
+                                    blank=True, null=True,
+                                    max_length=256, db_column='commune')
+
+    geom = models.PointField(verbose_name=_(u"Emplacement"), db_column='geom',
+                             blank=True, null=True,
+                             srid=settings.SRID, spatial_index=False)
 
     class Meta:
         db_table = 'o_b_renseignement'
@@ -698,6 +712,37 @@ class InformationDesk(models.Model):
 
     def __unicode__(self):
         return self.name
+
+    @property
+    def description_strip(self):
+        return strip_tags(self.description)
+
+    @property
+    def latitude(self):
+        if self.geom:
+            api_geom = self.geom.transform(settings.API_SRID, clone=True)
+            return api_geom.y
+        return None
+
+    @property
+    def longitude(self):
+        if self.geom:
+            api_geom = self.geom.transform(settings.API_SRID, clone=True)
+            return api_geom.x
+        return None
+
+    @property
+    def photo_url(self):
+        if not self.photo:
+            return None
+        thumbnailer = get_thumbnailer(self.photo)
+        try:
+            thumb_detail = thumbnailer.get_thumbnail(aliases.get('medium'))
+            thumb_url = os.path.join(settings.MEDIA_URL, thumb_detail.name)
+        except InvalidImageFormatError:
+            thumb_url = None
+            logger.error(_("Image %s invalid or missing from disk.") % self.photo)
+        return thumb_url
 
 
 class POIManager(models.GeoManager):